Things added as a result of compiling at a Solaris system; compiler pleasing, and one instance of link libraries updated.

+++ utils/vacation/vacation.c
@@ -224,7 +224,7 @@ main(argc, argv)
 	    usage();
 	  pw = getpwuid(getuid());
 	  if (!pw) {
-	    fprintf(stderr, "vacation: no such user uid %u.\n", getuid());
+	    fprintf(stderr, "vacation: no such user uid %ld.\n", (long)getuid());
 	    exit(EX_NOUSER);
 	  }
 	} else if (!(pw = getpwnam(*argv))) {
